Add a mainline arm64 device
The existing 'generic' targets are emulator devices, and targeting
it has the side-effect of including emulator-specific artifacts.

Create a mainline device instead, which is really generic. This also
allows specifying tweaks unsuitable for the emulator.

Include BoardConfigGsiCommon for now, even though it includes a few
things we likely won't want in the future.

Bug: 80410283
Bug: 111538404
Test: lunch mainline_arm64; m
Change-Id: I15e7cb41fcbd3c0592a091ba7a8575ddf10bda71