Update patch set 1
Patch Set 1:

Changing the names of the libs on the boot classpath may break a few things that "know" details of the boot classpath. e.g. vogar, but also sandboxing apps, etc.

I'm not thinking about core-simple, which is mostly for testing and we'll remove ASAP, but if your plan is to do the others (core-oj/core-libart, ext, framework, etc.) we (runtime, core lib, you) should talk about whether the naming change is worth it. If we could rename the build rule output to the original names (perhaps as an additional build step) and not affect the device details at all that would be safest.

Patch-set: 1
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1038039 <1038039@85c56323-6fa9-3386-8a01-6480fb634889>