Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Changing the names of the libs on the boot classpath may break a few things that "know" details of the boot classpath. e.g. vogar, but also sandboxing apps, etc.
> 
> I'm not thinking about core-simple, which is mostly for testing and we'll remove ASAP, but if your plan is to do the others (core-oj/core-libart, ext, framework, etc.) we (runtime, core lib, you) should talk about whether the naming change is worth it. If we could rename the build rule output to the original names (perhaps as an additional build step) and not affect the device details at all that would be safest.

As in [1] and [2], we look at sdk_version and link with the appropriate stubs library. To do this linking we have a naming rule. The actual library is {module_name}.impl, {module_name}.stubs for public, and {module_name}.stubs.system for system.
I know it's dangerous to change the name to bootclasspath, but I think it's right to change the name to take advantage of the java_sdk_library.

[1] https://cs.corp.google.com/android/build/make/core/java_common.mk?rcl=ab752bb403f8412ab137d2c548fffe996b36af80&l=300
[2] https://cs.corp.google.com/android/build/soong/java/java.go?rcl=4e9d3608877d4de67a791b70386715c210e2c5ac&l=860

Patch-set: 1
Reviewer: Gerrit User 1187124 <1187124@85c56323-6fa9-3386-8a01-6480fb634889>