Support retrofitting dynamic partitions.
When PRODUCT_RETROFIT_DYNAMIC_PARTITIONS is set, the maximum of sum of sizes
for all partitions for a slot is BOARD_SUPER_PARTITION_SIZE, even if it is
an A/B device. So:

- A/B retrofit:     BOARD_SUPER_PARTITION_SIZE
- non A/B retrofit: BOARD_SUPER_PARTITION_SIZE
- A/B launch:       BOARD_SUPER_PARTITION_SIZE / 2
- non A/B launch:   BOARD_SUPER_PARTITION_SIZE

Also, don't build super.img for retrofit devices, since it is
useful for shipping a device only. (super.img includes all
images in its _a slot, and is meant to be flashed directly
on the block device super).

Test: builds with retrofit device
Bug: 116608795
Change-Id: I562211b8e0abb7264b6b8b7ccb42601bd4d0cdce