Update patch set 1
Patch Set 1:

I'm sorry, I don't understand why we need to change the name of the file because of a build change.

The build system is supposed to support what the platform needs, not the other way round.

I have no problem with java_sdk_library, I like it if it saves boilerplate, but this change seems very invasive (and won't stop with this one library). Can't we just have a "filename" or "output_name" property on java_sdk_library that allows us to override the default jar name? i.e. instead of outputting foo.impl.jar, when filename is specified it uses the value and allows us to stop the artifact changing where it affects things on device?

Patch-set: 1
Reviewer: Gerrit User 1038039 <1038039@85c56323-6fa9-3386-8a01-6480fb634889>