WIP: Build hashtree for verified boot 2.0
The current approach is to dump the hashtree_info in plain text from
the avbtools; while the build scripts parse this information and try
to reconstruct the hashtree.

The plain text part can be improved if we can import the avbtools as
python libraries; or has better support for python based protobuf
generation so that avbtools can write the protobuf output directly.

An alternation approach is to parse the avb signed image from scrach,
and here would be the steps involved:
1. Unpack the AVB footer and find the vbmeta offset
2. Read the vbmeta header and find the descriptor offset
3. Iterate over the descriptors and find the one for hashtree
4. Unpack the Hashtree descriptor and dump the information.

Bug: 110100010
Test: unittest pass
Change-Id: I10db7095a45ea3dc6dba23a2067779514cfec709