build: improve first pass estimate for right sized images
Neglected to correct for filesystem size in estimate for verity.
Reinstate original heuristic for calculating the space freed by the
reduction in free space management, but use _change_ in size rather
than free_size for the math.  Recognize for debug situations without
verity signing that our estimate does need to be less aggressive.

To keep build space optimized and make sure that the existing image
does not confuse the tools, remove the out_file after the test build
image phase.

Test: build
Bug: 111302946
Change-Id: I360775d862d02a721d061bbc43e8b4e2f03bd01e