Add TARGET_FLATTEN_APEXS
TARGET_FLATTEN_APEXS, when set to true, flattens APEXs that are on
/system. Flattenig an APEX means files in it are not packaged as a mini
file system image, but exist in the /system partition as individual
files. This option is for devices where kerel does not support loopback
devices or the maximum number of loopback devices is too small (though
the threshold is TBD as of now).

This option must be turned off (i.e. APEXs must not be flattened) for
the devices launching with Q and beyond.

Bug: 118485880
Test: TARGET_FLATTEN_APEXS=true m apex.test; tree
out/target/product/.../system/apex/apex.test shows list of files in it.

Test; m apex.test, then a file out/target/product/.../system/apex/apex
.test.apex exists.

Change-Id: Ie13cb062c1387d55689692ba2b123e606ada8aa4