Set config.disable_cameraservice=0 by default.
I plan to reference this from an initrc to determine whether
cameraserver should be started; for this to work with "on property..."
the property needs a default value. 0 is already used as default in
the codebase.

Adding it in base_system.mk since that's where the 'cameraserver'
package is added. 'enabled' would be nicer, but the property exists
already in other places.

Bug: 117139498
Bug: 118369095
Test: as part of cameraserver.rc change I20cc7a956c561517b340d8b69382de713dda74e3
Change-Id: I9cce3e4dc2b57339cdc44830e19aa0218b8282c3