Update patch set 1
Patch Set 1: Code-Review+1

llkd can spike with a lot of existing problem(s) that need to be addressed immediately in the kernel. It is _first_ assumed that the kernel has at least passed lockdep, and that there are no kernel (driver) threads persistently hanging around in D state (can be defocussed/mitigated by ro.khungtask.enable=false to tell khungtask to stop, but one should really _fix_ the problems in the kernel drivers).

Turn it on, but be prepared to field bugs (some of which may stop up TreeHugger).

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1