Honor `PRODUCT_ART_TARGET_INCLUDE_DEBUG_BUILD` in Runtime APEX selection.
Include the Release Runtime APEX module on user builds and the "Debug"
Runtime APEX module on userdebug and eng builds, unless
`PRODUCT_ART_TARGET_INCLUDE_DEBUG_BUILD` is set. If so, and if this
variable is `true` (e.g. on marlin and sailfish), include the
Release (smaller) Runtime APEX module; if it is `false`, include the
"Debug" (larger) Runtime APEX module.

This logic is similar to the one used for module `art-runtime`
regarding debug variants in art/Android.mk.

Test: make systemimage on marlin-userdebug
Test: make systemimage on sailfish-userdebug
Test: make systemimage on marlin-eng
Bug: 113373927
Change-Id: I8adf605dd24dcc929359ac2b46a2f83a42f2ab8d