Honor `PRODUCT_ART_TARGET_INCLUDE_DEBUG_BUILD` in Runtime APEX selection.
Include the Release Runtime APEX module on user builds and the "Debug"
Runtime APEX module on userdebug and eng builds, unless
`PRODUCT_ART_TARGET_INCLUDE_DEBUG_BUILD` is set. If so, and if this
variable is `true`, include the Release (smaller) Runtime APEX module;
if it is `false`, include the "Debug" (larger) Runtime APEX module.

This logic is similar to the one used for module `art-runtime`
regarding debug variants in art/Android.mk.

Continue to honor `DONT_INCLUDE_RUNTIME_APEX` for now: if that
variable is `true`, don't include any Runtime APEX module.

Test: make systemimage with user target.
Test: make systemimage with userdebug/eng target.
Bug: 113373927
Change-Id: I8adf605dd24dcc929359ac2b46a2f83a42f2ab8d