Update patch set 2
Patch Set 2:

This is probably fine if it works.

This is super late to unmount product though.  Devices should be mounting product in first stage mount.  If it's not unmounted all the way at post-fs, then some parts of product may be in use.  May be worth adding a separate umount /product in the early-init trigger too to capture this case?

There's no way that you're going to be able to prevent the init scripts from being loaded from product by unmounting product in the init scripts.  If that is a requirement, then you'll need to work around this in init itself.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>