Pass --copy-dex-files=false to dex2oat for LOCAL_UNCOMPRESS_DEX
Tell dex2oat explicitly that the dex files will be uncompressed
in the final module Instead of relying on dex2oat to find
uncompressed dex files in the input jar.  This allows always
creating the input jar compressed, which means it is no longer
product-specific and can be moved to the common intermediates
directory in the next patch, preventing the command line embedded
in the odex file from varying between products.

Bug: 119412419
Test: m checkbuild
Change-Id: I387692b18f575ae9287ed9fed81ef00baceb7e12