Remove the reference to dumpkey in the doc
We used to dump the public key in the .x509.pem certificate with a host
java program. The dumped key is then processed and verified against in
recovery.

Now we removed the intermediate format as well as the dumpkey tool. And
the recovery code reads the certificate directly from otacerts.zip,
similar to flow in normal boot.

Bug: 116655889
Test: mma
Change-Id: I395dc46651013685679ce6743f1ffda4bc8e261a