releasetools: Keep the original RangeSet for 'uses_shared_blocks' case.
'uses_shared_blocks' indicates a case that the block list for a given
file is incomplete, where some of the blocks are "owned" by others. As
a result, such a file will be skipped from imgdiff'ing.

This CL makes a copy of the original block list before removing the
shared blocks. It uses the original RangeSet as the value of
extra['uses_shared_blocks']. validate_target_files.py will try to read
the original RangeSet, unless the original list is also incomplete or
has non-monotonic ranges.

Test: Run validate_target_files on a target that uses
      `BOARD_EXT4_SHARE_DUP_BLOCKS := true`.
Change-Id: I259e871ecc249ba0c14b5796bef413185a1b8242