Update patch set 7
Patch Set 7:

I am a bit confused as to why we have a mostly complete ELF parser for checking things that could instead be using a combination of readelf/nm/objdump. Do we feel like the formatting of those tools is variable enough that we couldn't capture DT_NEEDED as well as determining defined/needed symbols? This is the first time I heard we were going to do this extra validation (which is awesome), but I think the design might be a bit more complex than it needs to be.

Patch-set: 7
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>