blockimgdiff: selectively convert 'diff' commands to 'new' to reduce stash size
We cannot simultaneously stash more blocks than the size limit imposed by
the cache size. As a result, some 'diff' commands will be inevitably
converted to new. We used to do this conversion blindly when iterating
through the transfer list. This leads to an unintended large package.

In order to choose the right transfers to convert, we calculate the size
of the compressed data, and build a heuristic about the package size
increase to remove each stash blocks. After the process, the given
package size for shiner reduces from 194M->155M.

Bug: 120561199
Test: build incrementals on shiner and swordfish with customed threshold
Change-Id: I599420a91b80f1a1d83d22ee1b336b699050cfb4