Update patch set 1
Patch Set 1: Code-Review-1

Treble defines that a OEM, ODM and VENDOR as separate layers of manufacturing each with their responsibilities. Not using it in Pixel's few devices does not mean that the 13000 devices out there are not, or will not in some future, use it.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1