Use zopflipng to compress the recovery background text images
This saves space in the boot image at the cost of the increase of build
time. In practice, it adds about ~20 seconds to build the bootimage with
32 threads.

zopflipng generally runs slower. Even with a predefined filters
strategy, it runs about 3 times slower than pngcrush: ~40s vs 15s to
compress a 1.8M png file. However, zopflipng also gives a better
compression: ~1.4M vs 1.6M for the same image. So I guess the time
consumption is a price worth paying.

Bug: 74397117
Test: build the bootimage and check the resource images
Change-Id: I3e71f8bedd3c183a3374b1a9464bf05500e1aa61