Update patch set 2
Patch Set 2:

I'll try to get my CL in tomorrow, but it might be hard to get reviews since Dan is out.  I'd suggest redoing this change on top of my change, since that work is going to have to be done regardless, and then if there is any delay getting my change in we can submit this one and squash the new version into my change.

Patch-set: 2