Apply skip mounting product and other partitions in GSI
Uses skip_mount.cfg for `init` to skip mounting product and
product_service paritions in GSI. The patch also removes
the un-mount in `init.gsi.rc`.

With this patch, a device with flashing GSI will not use the
content in product/prodcut_service partitions on the device
and will always use the content under /system/product in GSI.

The patch also rename the target/product/vndk to
target/product/gsi.

Bug: 120208657
Bug: 117246558
Test: Boot blueline master/P with flashing aosp_arm64-userdebug
Test: /product and /product_service should be symbolic links
Change-Id: Idc9c60e5a05b3d90fc4ab64b0ac000571849b19b