Configuration the OTA package public key through PRODUCT_OTA_PUBLIC_KEYS
The recovery will verify the legitimacy of the 'ota package' when
the devices upgrade through the recovery. But the default ota public
keys may not be the final keys signing the ota package. The product
can configuration the OTA package public key through PRODUCT_OTA_PUBLIC_KEYS,
so the rules with verify the ota keys in the recovery image can work fine.

Test: make incremental ota package and check the ota package upgrade successfully
Change-Id: I40fb19f7c309955dd36cc73bfbfd55c83a12e189