Add build machinery for SCS vDSO workaround to build/make.
Treble requires us to be able to run if the kernel is from an older
release than the system partition. This means that it's possible for
us to be running on a device with a kernel whose vDSO clobbers the
x18 register (which SCS requires to be preserved) because [1] caused
the vDSO in the Pixel 2 and 3 kernels released with P to clobber
x18. This patch never landed in the upstream or common kernels, so
(to the best of my knowledge) only Pixel 2 and 3 are affected. A newer
version of the patch that is currently being considered for upstream
and common includes a fix for this issue [2]. The fix is also present
in the development versions of the Pixel 2 and 3 kernels.

This patch contains the build machinery that supports being able to
activate the bionic code that preserves x18 on a per-device basis,
and also activates it in the GSI unconditionally because we don't
know whether the GSI will be flashed onto a device that needs the
workaround.

[1] https://android.googlesource.com/kernel/msm/+/1bfa5df2a35bb81cd58f6c87cf3695ed880fac50
[2] https://patchwork.kernel.org/patch/10671685/

Bug: 122331452
Test: enabled SCS in libbluetooth{,-binder,_jni}.so and built aosp_walleye-userdebug, device still boots and Bluetooth works
Change-Id: I41f9a87b85fc0db276c62a0e66d43170e7af5b31