Update patch set 1
Patch Set 1:

Are these changes going to be reverted once we bring in more stuff from the internal tree? Are they just so we get a green build for now until the soong changes merge in to allow things to actual build?

Patch-set: 1
Reviewer: Gerrit User 1052976 <1052976@85c56323-6fa9-3386-8a01-6480fb634889>