Update patch set 1
Patch Set 1:

This only adds it for the legacy devices right? I don't understand the build makefiles very well, but, what happens with the first stage ramdisk build and this changes? Since we already have the ramdisk back, I'd like to make sure this is equivalent to including these modules
in the ramdisk as well.

Note that we are not including these modules in ramdisk (yet), but eventually we may.

cc: tomcherry@ and smuckle@ who would know better than me in this case.

Patch-set: 1
Reviewer: Gerrit User 1108565 <1108565@85c56323-6fa9-3386-8a01-6480fb634889>