Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1: Code-Review+1
> 
> (1 comment)
> 
> nice. any diffs for products not using prebuilts?

There weren't any files added or removed from any of the treehugger tested builds, so it doesn't seem so.

The only intermediates it changed on aosp-master/aosp_arm64 were fake_packages (which are fine, we evaluated all their required modules into modules_to_install) and one of the files in recovery, but that was just because we're not building a recovery image for that configuration.

Patch-set: 1
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>