Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-2
> > 
> > Use BUILD_PREBUILT when BUILD_MULTI_PREBUILT does not handle the options you need.
> 
> Our Android.mk is like this:
> include $(CLEAR_VARS)
> 
> LOCAL_PREBUILT_LIBS := libCameraVerfication:libs/armeabi-v7a/libCameraVerfication.so
> LOCAL_PREBUILT_LIBS += libopencv_java3:libs/armeabi-v7a/libopencv_java3.so
> 
> include $(BUILD_MULTI_PREBUILT)
> 
> if we replace BUILD_MULTI_PREBUILT with BUILD_PREBUILT, built failed.
> error log is below:
> 
> build/make/core/prebuilt_internal.mk:10: error: dont use LOCAL_PREBUILT_LIBS anymore LOCAL_PATH=vendor/sprd/platform/packages/apps/ValidationTools.
> 08:47:20 ckati failed with: exit status 1

Yes, it's not a direct replacement, as BUILD_MULTI_PREBUILT is a shortcut. It's translating to roughly this:

  include $(CLEAR_VARS)
  LOCAL_MODULE := libCameraVerification
  LOCAL_MODULE_CLASS := SHARED_LIBRARIES
  LOCAL_SRC_FILES := libs/armeabi-v7a/libCameraVerification.so
  LOCAL_MODULE_SUFFIX := .so
  include $(BUILD_PREBUILT)

  include $(CLEAR_VARS)
  LOCAL_MODULE := libopencv_java3
  LOCAL_MODULE_CLASS := SHARED_LIBRARIES
  LOCAL_SRC_FILES := libs/armeabi-v7a/libopencv_java3.so
  LOCAL_MODULE_SUFFIX := .so
  include $(BUILD_PREBUILT)

Which you can then add LOCAL_PRODUCT_MODULE to.

That being said, that looks like the directories for libraries used by a prebuilt APK? You can just keep the .so files embedded inside the APK. Just make sure that they are uncompressed in the installed APK (STOR when using zipinfo). The build system should be doing that for you, but it can be turned off.

Patch-set: 1
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>