Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Note that ODM can also be a dynamic partition, and it is very small. (Pixel 3 doesn't have a real ODM partition, but the tree size of /vendor/odm is 32K).
> 
> I have strong hopes _this_ change will have no problems with it.
> 
> /product_services and some system_other partitions are almost completely empty, and those built. Any problems we have in these set will solve ODM partition problems, hopefully we will get more examples of DAP products as we move forward to prove this out.

Shall we put some testcases into test_build_image.py that generate formerly-failing images (e.g. if we can mimic the image size, or the number of files to go into the image)? It'd be helpful for tuning the parameters in future. Internally we have tests that run test_build_image.py continuously (go/otatest), so we'll also receive signals in case the breakage comes from elsewhere.

Patch-set: 1