Add whitelists for LEGACY_HIDL_APPS.
wrt libs:
android.hidl.base-V1.0-java
android.hidl.manager-V1.0-java

These used to be on the boot classpath, but are half published because
of 1414929, and so the libs were removed in 86259915. However, because
they are added to all apps (and only a relatively small number of apps
use these libraries), this causes a ~21ms regression in <= P targeting
apps' startup times.

Bug: 86259915
Test: manually adding entries to PRODUCT_*_LEGACY_HIDL_APPS and
    checking output xmls.
Test: boot, sanity check logs

Change-Id: Ic8204b4de300f6efbe2172374e4272e453cfb003