Skip some build steps if not building system image.
core/Makefile: This patch also excludes building of the OTA package if we're
not building a system image. In this case the OTA package would be generated as
a later step.

tools/releasetools/add_img_to_target_files.py: This patch also excludes the
inclusion of the system path into the target files if it does not exist (which
it will not if we are not building the system image).

Bug: 123430711
Test: Disable building system image, verify that target files builds, OTA does not.
Change-Id: Iaf964ede2b1df5ea4e004b572fd91187a366e75e