First pass at creating PRODUCT_HOST_PACKAGES
This only defines PRODUCT_HOST_PACKAGES as a concept, and requires that
it contain at least all of the files that would be installed to the host
from PRODUCT_PACKAGES (minus libraries, as those are expected to be
installed based on what's required by other modules).

It does not actually change what will be built, that will happen in a
future change.

Test: build/soong/build_test.bash
Change-Id: Ie502b1a483f821b957cfd407c31a93e243896dae
Merged-In: Ie502b1a483f821b957cfd407c31a93e243896dae