Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > nfuller@ told me that reverting CleanSpec is not a good idea, but I don't remember the details.
> 
> Sorry, I mean CleanSpec is for incremental build. If you are moving artifacts in this revert, you may need to add something in the CleanSpec, rather than remove

My understanding (partly assumption / guesswork / empirically arrived at) is that CleanSpec entries work by being recognized by index. So, I think it's additive only.

e.g.
if you add entry [5] and it is encountered by a build, the high watermark for that build is 5
if you revert entry [5], the high watermark for that build is still 5.
if you add a different/new entry [5] it won't be run on devices that have previously run the old [5]

I *could* be wrong.

Patch-set: 2