Update patch set 2
Patch Set 2:

(2 comments)

I have some drive-by comments.  I've marked them "resolved" since my intent isn't to hold up this CL.

In my naive mind, reducing the number of shell subprocesses we invoke is worthwhile on something which gets run on every build on the build server, since we run so many builds and always seem waiting for builds to complete.  But what I'm noting might be on the order of  0.0001% or something, and so not worth thinking about.

Thanks!

Patch-set: 2