Add TARGET_SKIP_ADD_IMG_TO_TARGET_FILES for partial target files packages.
This change is an optimization to the build process when generating a partial
build. The change avoids adding system images that won't be used, and defers
that operation until a later step once all of the images are assembled from the
various partial builds that contribute to the merged build.

When building a partial target files package that will be merged with another
partial target files package later (using merge_target_files.py), the merge
step will add the images to the merged target files package, so we can skip
that step in the partial builds. This change introduces a new make variable,
TARGET_SKIP_ADD_IMG_TO_TARGET_FILES, that allows skipping that step.

Test: Regular build: no change
Test: Partial builds: no images in generated partial target files
Bug: 122637948
Change-Id: I70a84805941402c6da29b360e6a9714f3043d4bd